+++ common/src/main/java/cc/cassian/pyrite/Pyrite.java
@@ -122,6 +122,12 @@ public class Pyrite {
 			generateBrickSet(dye+"_terracotta_brick", Blocks.TERRACOTTA, color, blockLux);
 			// Dyed Torches
 			createTorch(dye+"_torch", getTorchParticle(dye));
+			if (dyeIndex > 15) {
+				// Dyed Stained Glass
+				createPyriteBlock(dye+"_stained_glass","stained_framed_glass", 0.3f, color, blockLux);
+				// Dyed Stained Glass Pane
+				createPyriteBlock(dye+"_stained_glass_pane","stained_framed_glass_pane", 0.3f, color, blockLux);
+			}
 			// Dyed Lamps
 			createPyriteBlock(dye + "_lamp","block", 0.3f, color, 15);
 			// Dyed Framed Glass